fixed bug and changed role permissions syntax

+++ src/main/java/com/rata/userService/records/newRecords/RoleSearchCriteria.java
@@ -1,7 +1,7 @@
 package com.rata.userService.records.newRecords;
 
 public record RoleSearchCriteria(
-        Long appId,
+        Integer appId,
         String search,
         Boolean systemRole,
         Boolean status
